ai-medical-prescription/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pydantic
python-multipart
aiofiles
python-dotenv
httpx
plotly
pandas
